third_party/WebKit/LayoutTests/webaudio/osc-sawtooth-sweep-snr-expected.txt

Test Sawtooth Oscillator with Exponential Sweep

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 83 dB
PASS Maximum difference below threshold of 3.18 ulp (16-bits)
PASS Number of differences between actual and expected result is less than 4100
PASS successfullyParsed is true

TEST COMPLETE
